Introduction of some wrappers/decorators.
This patch fix #ateliedocodigo/py-healthcheck#21

What has been done?

1. Added the healthcheck/wrappers.py file/submodule.
2. Added the `wrappers.add_check_to` wrapper.
3. Added the `wrappers.add_section_to` wrapper.
4. Increased the version number into `setup.py`.

About `wrappers.add_check_to`:

As I do not want to create a global variable, or a unify the way to
reference to the `HealthCheck` class, I made it clear that it is needed
as reference. If it's not the intended behavior, let me know.

  Note: It raises an exception if it get an input class which is not
  `HealthCheck`.

About `wrappers.add_section_to`:

As I do not want to create a global variable, or unify the way we
reference to `HealthCheck` or `EnvironmentDump`, I made it clear that
one of them is needed. The second argument is the name of the section
(nothing new).

  Note: It raises an exception if it get an input class which is not
  `HealthCheck` or `EnvironmentDump`.